# api/v1alpha1/openapi.yaml
openapi: 3.0.4
info:
  title: 3-Tier Demo Service Provider API
  description: |
    DCM service provider for 3-tier (web, app, db) demo apps.
    AEP-compliant REST API.
  version: 1.0.0
  contact:
    name: DCM Project
    url: https://github.com/dcm-project/3-tier-demo-service-provider
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html

servers:
  - url: /
    description: 3-Tier Demo Service Provider

tags:
  - name: Health
    description: Health check
  - name: ThreeTierApps
    description: 3-tier app lifecycle

paths:
  /api/v1alpha1/three-tier-apps/health:
    get:
      summary: Health Check
      description: Returns the health status of the service
      operationId: getHealth
      tags:
        - Health
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Health'

  /api/v1alpha1/three-tier-apps:
    get:
      summary: List Three-Tier Apps
      description: List 3-tier demo apps with optional pagination
      operationId: listThreeTierApps
      tags:
        - ThreeTierApps
      parameters:
        - name: max_page_size
          in: query
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 50
        - name: page_token
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThreeTierAppList'
        '400':
          description: Bad request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Create Three-Tier App
      description: Create a new 3-tier demo app (DB, app, web)
      operationId: createThreeTierApp
      tags:
        - ThreeTierApps
      parameters:
        - name: id
          in: query
          description: |
            Optional client-specified ID for the three-tier app. If not provided,
            the server will generate a UUID. When both metadata.name and id are set,
            id is the stored instance id.

            Requirements (per AEP-122):
            - 1-63 characters long
            - Start with a lowercase letter or digit
            - Contain only lowercase letters, numbers, and hyphens
            - End with letter or number
          required: false
          schema:
            type: string
            pattern: '^[a-z0-9]([a-z0-9-]{0,61}[a-z0-9])?$'
            minLength: 1
            maxLength: 63
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ThreeTierApp'
      responses:
        '201':
          description: Three-tier app created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThreeTierApp'
        '400':
          description: Bad request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/v1alpha1/three-tier-apps/{threeTierAppId}:
    get:
      summary: Get Three-Tier App
      description: Get a 3-tier app by ID
      operationId: getThreeTierApp
      tags:
        - ThreeTierApps
      parameters:
        - name: threeTierAppId
          in: path
          required: true
          schema:
            type: string
            pattern: '^[a-z0-9]([a-z0-9-]{0,61}[a-z0-9])?$'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThreeTierApp'
        '404':
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete Three-Tier App
      description: Delete a 3-tier app
      operationId: deleteThreeTierApp
      tags:
        - ThreeTierApps
      parameters:
        - name: threeTierAppId
          in: path
          required: true
          schema:
            type: string
            pattern: '^[a-z0-9]([a-z0-9-]{0,61}[a-z0-9])?$'
      responses:
        '204':
          description: Deleted
        '404':
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    Health:
      type: object
      description: Health status
      x-aep-resource:
        singular: health
        plural: healths
        singleton: true
        type: three-tier-demo-sp.dcm.io/health
        patterns:
          - health
      required:
        - type
        - state
        - path
      properties:
        type:
          type: string
          format: uri-reference
          readOnly: true
          example: "three-tier-demo-sp.dcm.io/health"
        state:
          type: string
          example: "healthy"
        path:
          type: string
          readOnly: true
          example: "health"

    ThreeTierApp:
      type: object
      description: 3-tier app (web, app, db). Create requires spec; id from query param, metadata.name, or server UUID.
      x-aep-resource:
        type: three-tier-demo-sp.dcm.io/three-tier-app
        singular: three-tier-app
        plural: three-tier-apps
        patterns:
          - three-tier-apps/{three_tier_app_id}
      required:
        - spec
      properties:
        id:
          type: string
          readOnly: true
          pattern: '^[a-z0-9]([a-z0-9-]{0,61}[a-z0-9])?$'
          description: Unique app ID (server-generated from metadata.name)
        path:
          type: string
          readOnly: true
          description: Resource path (server-generated)
        metadata:
          $ref: '#/components/schemas/ThreeTierAppMetadata'
        spec:
          $ref: '#/components/schemas/ThreeTierSpec'
        status:
          readOnly: true
          allOf:
            - $ref: '#/components/schemas/ThreeTierAppStatus'
        service_type:
          type: string
          enum: [three_tier_app_demo]
          description: Service type (must be three_tier_app_demo)
        provider_hints:
          type: object
          additionalProperties: true
        web_endpoint:
          type: string
          readOnly: true
          description: |
            URL to reach the web tier, populated when an external IP is available
            (e.g. on OpenShift with a LoadBalancer). Nil when the cluster does not
            assign an external IP (use `kubectl port-forward` instead).
        create_time:
          type: string
          format: date-time
          readOnly: true
        update_time:
          type: string
          format: date-time
          readOnly: true

    ThreeTierAppMetadata:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          pattern: '^[a-z0-9]([-a-z0-9]*[a-z0-9])?$'
          minLength: 1
          maxLength: 63
          example: my-petclinic
        labels:
          type: object
          additionalProperties:
            type: string

    ThreeTierAppStatus:
      type: string
      enum: [PENDING, RUNNING, FAILED, DELETED]
      description: Aggregated status of the 3-tier app

    ThreeTierAppList:
      type: object
      properties:
        three_tier_apps:
          type: array
          items:
            $ref: '#/components/schemas/ThreeTierApp'
        next_page_token:
          type: string

    ThreeTierSpec:
      type: object
      description: Three-tier structure (database, app, web)
      required:
        - database
        - app
        - web
      properties:
        database:
          $ref: '#/components/schemas/DatabaseTierSpec'
        app:
          $ref: '#/components/schemas/AppTierSpec'
        web:
          $ref: '#/components/schemas/WebTierSpec'

    DatabaseTierSpec:
      type: object
      description: |
        Database tier. Uses abstract identifiers (engine, version) only.
        SP maps engine+version to OCI image (e.g. postgres+16 -> docker.io/library/postgres:16).
      required:
        - engine
        - version
      properties:
        engine:
          type: string
          description: Database engine (postgres, mysql).
          example: postgres
        version:
          type: string
          description: Database engine version.
          example: "16"
        network:
          $ref: '#/components/schemas/TierNetwork'

    AppTierSpec:
      type: object
      description: Application tier. Uses OCI image.
      required:
        - image
      properties:
        image:
          type: string
          description: OCI image reference for the application.
          example: docker.io/springcommunity/spring-framework-petclinic:6.1.2
        http_port:
          type: integer
          minimum: 1
          maximum: 65535
          description: |
            Primary HTTP port the application listens on.
            Used by the web tier proxy to forward traffic. Defaults to 8080.
          example: 8080
        network:
          $ref: '#/components/schemas/TierNetwork'

    WebTierSpec:
      type: object
      description: Web tier. Uses OCI image.
      required:
        - image
      properties:
        image:
          type: string
          description: OCI image reference for the web server.
          example: docker.io/library/nginx:alpine
        network:
          $ref: '#/components/schemas/TierNetwork'

    TierNetwork:
      type: object
      properties:
        ports:
          type: array
          items:
            $ref: '#/components/schemas/ContainerPort'

    ContainerPort:
      type: object
      additionalProperties: true
      required:
        - container_port
        - visibility
      properties:
        container_port:
          type: integer
          minimum: 1
          maximum: 65535
          example: 8080
        visibility:
          type: string
          description: |
            How this port is exposed to consumers.
            - none: Port is not exposed outside the container process
            - internal: Exposed to the host or cluster network
            - external: Reachable from outside the host/cluster
          enum:
            - none
            - internal
            - external
          example: internal

    Error:
      type: object
      description: RFC 7807 problem details
      required:
        - type
        - title
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri-reference